requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-multipart>=0.0.6
sse-starlette>=2.0
python-frontmatter>=1.1
markdown>=3.6
argon2-cffi>=23.1
PyJWT>=2.8
requests>=2.31
readability-lxml>=0.8
beautifulsoup4>=4.12
yt-dlp>=2024.5
